výpis docházky a uzavření, schválení u jednorázové události

+++ trainings/forms.py
@@ -18,6 +18,7 @@ from events.forms_bases import (
     OccurrenceFormMixin,
     PersonMetaMixin,
     ActivePersonFormMixin,
+    ReopenOccurrenceMixin,
 )
 from events.models import (
     EventOrOccurrenceState,
@@ -851,7 +852,7 @@ class TrainingEnrollMyselfParticipantOccurrenceForm(
         return instance
 
 
-class FillAttendanceForm(ModelForm):
+class TrainingFillAttendanceForm(ModelForm):
     class Meta:
         model = TrainingOccurrence
         fields = []
@@ -952,12 +953,11 @@ class FillAttendanceForm(ModelForm):
                     entity_assignment.state = TrainingAttendance.PRESENT
                     if type(entity_assignment) is CoachOccurrenceAssignment:
                         occurrence_date = instance.datetime_start.date()
-                        rate = PersonHourlyRate.objects.filter(
-                            person=entity_assignment.person,
-                            event_type=instance.event.category,
-                        ).first()
-                        if rate is not None:
-                            hourly_rate = rate.hourly_rate
+                        person_rates = PersonHourlyRate.get_person_hourly_rates(
+                            entity_assignment.person
+                        )
+                        if instance.event.category in person_rates:
+                            hourly_rate = person_rates[instance.event.category]
                             if entity_assignment.transaction is None:
                                 entity_assignment.transaction = Transaction(
                                     amount=hourly_rate * instance.hours,
@@ -991,16 +991,10 @@ class FillAttendanceForm(ModelForm):
         return instance
 
 
-class ReopenTrainingOccurrenceForm(ModelForm):
+class ReopenTrainingOccurrenceForm(ReopenOccurrenceMixin, ModelForm):
     class Meta:
         model = TrainingOccurrence
         fields = []
-
-    def clean(self):
-        cleaned_data = super().clean()
-        if not self.instance.can_be_reopened:
-            self.add_error(None, "Tato událost nemůže být znovu otevřena")
-        return cleaned_data
 
     def save(self, commit=True):
         instance = super().save(False)
@@ -1010,8 +1004,9 @@ class ReopenTrainingOccurrenceForm(ModelForm):
             occurrence=instance, state=TrainingAttendance.PRESENT
         )
         for present_coach_assignment in present_coach_assignments:
-            present_coach_assignment.transaction.delete()
-            present_coach_assignment.transaction = None
+            if present_coach_assignment.transaction is not None:
+                present_coach_assignment.transaction.delete()
+                present_coach_assignment.transaction = None
 
         observed_unexcused_assignments = [
             TrainingParticipantAttendance.objects.filter(
